requirements.txt
faster-whisper>=1.1.0
numpy>=1.26.0
PyAudioWPatch>=0.2.12.8
webrtcvad-wheels>=2.0.14
